+++ bad6.js
@@ -37,10 +37,10 @@ Hooks.once("init", async () => {
 	foundry.documents.collections.Actors.unregisterSheet("core", foundry.appv1.sheets.ActorSheet);
 
 
-
+/*
 	CONFIG.BAD6 = BAD6;
 	CONFIG.INIT = true;
-
+*/
 	foundry.documents.collections.Actors.registerSheet("bizarre-adventures-d6", UserSheet, {
 		types: ["user"]
 		, makeDefault: true